rename cmn to client
