+++ src/mako/api/api.rs.mako
@@ -27,7 +27,7 @@ use std::mem;
 use std::thread::sleep;
 use std::time::Duration;
 
-use crate::cmn;
+use crate::client;
 
 // ##############
 // UTILITIES ###
@@ -57,7 +57,7 @@ pub struct ${hub_type}${ht_params} {
     _root_url: String,
 }
 
-impl<'a, ${', '.join(HUB_TYPE_PARAMETERS)}> cmn::Hub for ${hub_type}${ht_params} {}
+impl<'a, ${', '.join(HUB_TYPE_PARAMETERS)}> client::Hub for ${hub_type}${ht_params} {}
 
 impl<'a, ${', '.join(HUB_TYPE_PARAMETERS)}> ${hub_type}${ht_params}
     where  ${', '.join(hub_type_bounds())} {
