+++ src/mako/api/lib/lib.mako
@@ -192,7 +192,7 @@ ${self.hub_usage_example(c, rust_doc, fr=fr)}\
 
 ${'##'} Handling Errors
 
-All errors produced by the system are provided either as ${link('Result', 'cmn::Result')} enumeration as return value of
+All errors produced by the system are provided either as ${link('Result', 'client::Result')} enumeration as return value of
 the ${api.terms.action}() methods, or handed as possibly intermediate results to either the 
 ${link('Hub Delegate', delegate_url)}, or the ${link('Authenticator Delegate', urls.authenticator_delegate)}.
 
@@ -200,9 +200,9 @@ When delegates handle errors or intermediate values, they may have a chance to i
 makes the system potentially resilient to all kinds of errors.
 
 ${'##'} Uploads and Downloads
-If a method supports downloads, the response body, which is part of the ${link('Result', 'cmn::Result')}, should be
+If a method supports downloads, the response body, which is part of the ${link('Result', 'client::Result')}, should be
 read by you to obtain the media.
-If such a method also supports a ${link('Response Result', 'cmn::ResponseResult')}, it will return that by default.
+If such a method also supports a ${link('Response Result', 'client::ResponseResult')}, it will return that by default.
 You can see it as meta-data for the actual media. To trigger a media download, you will have to set up the builder by making
 this call: `${ADD_PARAM_MEDIA_EXAMPLE}`.
