+++ src/rust/cli/client.rs
