+++ src/rust/cli/mod.rs
@@ -1,1 +1,1 @@
-pub mod cmn;
+pub mod client;
